+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './enums';
 export * from './interfaces';
 export * from './liquid-commerce-client';
+export * from './liquid-commerce-order-client';
 export * from './types';